Document manual-copying install method
This is not an ideal user experience. But this allows you to generate
the files necessary for system-wide installation as a regular user,
then install those files as root.  This way you are only invoking root
powers for cp(1) and you don't need to install and run xkalamine (and
its dependencies) as root.